Fix DAT connect
Previously, connect would throw if the agent version was not known. We
need to know the agent version to ensure that keysplitting is enabled before
performing keysplitting handshake. We also need the agentId to do a
valid keysplitting SYN.

But in the case of DAT targets, the agent version (and agentId, etc.) is
not known until the target is provisioned and calls back to the bastion.
With new connection-service changes, a bastion poller will now trigger
an API call on the connection node which will send a Ready event when
the DAT target is fully provisioned and agent version is known.

So instead of throwing with an unknown agent version error in the case
of DAT, a flag is set so that when receiving the Ready event the target
info will be refreshed.

+ Add DAT connect example